refactor(audio): make an extra effort for I/O devices  to be in logs
There are some scenarios where I/O info is still absent from logs. This
may be correct in some cases, but since we're not guaranteeing the list
of devices is always populated in audio-manager, we can't be sure.

Make an extra effort to guarantee that I/O devices are always logged in
audio failure entries. Guarantee they're enumerated before logging for
scenarios such as skipCheck + audio join failure.