fix(audio): muteOnStart state inconsistency with LiveKit
There's an issue with LiveKit-based audio where muteOnStart=true
environments have a state inconsistency on room join. Mute state would
be flagged as `muted`, but the actual mute state in LiveKIt would differ
to `unmuted` (i.e.: users would be heard while unmuted). This is partly
due to a bbb-webrtc-sfu issue where it didn't notify track
publish/unpublishes + the client mistakengly publishing tracks when it
shouldn't.

Adjust the client to actually use the muteOnStart state prior to joining
audio - and if true, do not publish mic tracks.
The webrtc-sfu part will be bumped in a subsequent commit.