fix(audio): autoplay handling for LiveKit
Autoplay handling is intermittent when using the LiveKit bridge since we
are trying to reuse the existing triggers and handlers (i.e.: the media
element one in audio-manager). Those are not suitable for all scenarios
in LiveKit, which may cause users to be unable to hear the whole room or
specific participants if autoplay heuristics kick in.

Make autoplay handling more reliable when using audio via LiveKit by
using their playback status events and playback handler rather than
relying on our custom handlers.