fix(livekit): add shutdown timeout to livekit-server/sip
LiveKit has a graceful shutdown procedure where it waits for all users
to stop. That's not the behavior of the rest of BBB components, nor is
it necessary for a regular BBB installation where LiveKit is contained
to the box and not scaled horizontally.

Add a timeout (5s) to LiveKit's services so that the graceful shutdown
is bypassed.