Improve getJoinUrl, supporting parameters such as enforceLayout and userdata-*
Now the client will read the enforceLayout of the session instead of reading it from the user.
And userdata will merge the session params with the user params.

The new params are required mainly for hybrid rooms, where the user that will show the meeting for the present students will clone the current user passing different params like enforceLayout=PRESENTATION_ONLY for the new session.