fix(webcam): add option to skip initial enumerateDevices
See: https://github.com/bigbluebutton/bigbluebutton/issues/21746
The client currently does an enumerate -> optional gUM -> enumerate flow
(the last enum only if necessary - i.e.: fingerprinting measures were in
effect in the first enumerate), which is an attempt to save extra
enumerate/gUM calls whenever necessary. The fact that it's breaking gUM
in Firefox seems like a browser bug, but we should work around it.

Add a flag (public.media.skipInitialCamEnumeration) that allows skipping
the initial optimistic enumerateDevices call to try to work around the
aforementioned issue. After this has been tested and validated, we can
port it to 3.0 without a feature flag behind it.