Improve scalability by running multiple hasura instances
At 20th BBB dev summit we ran some load tests on large meetings. It
turned out that Hasura is a bottleneck if the meeting has more than 400
participants. As hasura is stateless it can be scaled by running
multiple instances of it and run `nginx` as loadbalancer in front of it.

The request flow for `/graphql` is then:

* Browser → nginx (443)
* → bbb-graphql-middleware (8378)
* → nginx hasura loadbalancer (8185) (this is new)
* → hasura (8085, 8086, …)

The systemd unit file setup is designed in a way that for the hasura
instance an instance of the template unit `bbb-graphql-server@.service`
is created. These bind to `bbb-graphql-server.service` so that
`systemctl {start,stop} bbb-graphql-server.service` will start/stop all
registered instances.

`bbb-graphql-server.service` has a dependency to
`bbb-graphql-server@8085.service` so that there will always exist at
least one hasura instance.