BBB3 chat features for video recording format
The chat panel in the video recording format is restyled to look more
like the live BigBlueButton session, with timestamps shown beside the
user names and chat bubbles.

Emoji reactions are added, and so are replies (which show the first
line of the message being replied to). The user's name is only shown
for the first message sent in a series of messages by the same user.

The entire Popcorn plugin for chattimeline is basically rewritten, in a
more modern JS style, and to build the more complex HTML elements needed
for the new chat features.

The video format recording processing script is updated to add some
additional required attributes to the chattimeline in the video.xml
file. Note that because this is processed by Popcorn, the XML
structure is somewhat limited - nested elements like the presentation
recording format uses are not permitted. As such, the emoji reactions
are encoded as JSON within an attribute.