fix(audio): spurious talking events if LiveKit audio unused
A spurious talking event is sent from the LiveKit bridge when it's not
being used, but other bridges (eg camera or screen share)  area. This is
mostly benign as those events are falsy, but it introduces duplicate
entries in recording events.

Do not send talking events from the LiveKit bridge if it's not being
used for audio.