fix(audio): set inputDeviceId as listen-only if legacy LO is used
When the "legacy" listen only path is used (i.e.: via the modal or
lock settings), the input device ID is not being set to the
"listen-only" placeholder value. This causes two inconsistencies:
  - Default bridge: unmute action might not do what's expected
    (prompting the user for permission, opening the audio settings
    modal)
  - LiveKit bridge: the above + the user might be unable to speak unless
    they manually toggle between input devices in the live I/O device
    selector

Properly set the input device ID as "listen-only" whenever
AudioManager's joinListenOnly method is called and the bridge supports
transparent listen only. This addresses the issues mentioned above.