feat(livekit): selective subscription in camera bridge
Use selective subscription in LiveKit's camera bridge if enabled to do
so. This should only subscribe to streams when necessary now, which is more
efficient than the autosubscription approach that was in effect.

Additionally, multiple assorted fixes to camera-related features, e.g.
(not exhaustive):
  - Client might get stuck when a share attempt fails
  - Reduce spurious re-renders in the bridge component
  - Properly cleanup camera bridge streams on unmount

Config flag is public.media.livekit.selectiveSubscription. Disabled by
default for now.