fix(audio): client state not fully in sync when using LiveKit audio
The client state (i.e.: audio-manager's isConnected prop) might be
out-of-sync compared to the server state when using LiveKit. Sample
scenario: autoJoin=false or closing the audio modal.
Not joining audio is not an option with LiveKit as it happens automatically,
but audio-manager still presumes it is. This means the audio join callback
is only called when join is explictly initiated by one of it's joinAudio
calls, which causes the state desync.

Add tracking for the server audio state when using LiveKit and guarantee
audio-manager is in sync. This is done in a way that does not interfere
with other audio bridges.