fix(livekit): guarantee audio-manager inputStream is set
AudioManager's inputStream is not always set when using LiveKit (e.g.:
when skipping the audio-settings modal). This means the bridge and
manager's inputStream are desync, which may cause failures in components
such as MutedAlert which rely on the manager's state.

Guarantee AudioManager's state is always up to date with LiveKit's via:
  - Adding a new callback event to indicate that a stream was published
    (and hence may need to be set)
  - Guaranteeing the original input stream is return from the bridge's
    getter if no audio was published yet.