fix(whiteboard): Add timeout for ignoring presentation restoration on new events (#21668)
When the client is starting up the whiteboard seems to be doing some kind of zoom initialization which is affecting some effects of the presentation component. A timeout was added to ignore presentation restoration in the first 5 seconds after client startup. This fixes a bug where the config bbbHidePresentationOnJoin does not take effect.