fix(livekit): disable autoSubscribe if selectiveSubscription is enabled
Having autoSubscribe on even if selectiveSubscription is enabled is
causing issues on reload/re-connect scenarios where streams are being
subscribed to, but the selective subscription hook is not clearing them
up (it doesn't track external subscriptions).

Disable autoSubscribe if selectiveSubscription is on. In the future,
we'll adjust the selective subscription hook to watch for and clean up
unwarrented subscriptions.