Speed up upload of presentation by removing a fixed delay
BBB-Web used to wait 5 seconds for presentation uploads to make sure the meeting is already created.
This PR changes this behavior by adding an attempt each 1 seconds to check if the meeting is running.
It will improve a lot the speed, mainly for uploads during the meeting where the meeting is already created and it doesn't need any delay.