fix(sidebar-navigation): floating icons on page load
There was an issue where the icons from the navigation sidebar are
rendered before its dimensions were calculated by the layout context, so
the icons kept floating in the UI during page load. The fix is just a
check to prevent the sidebar navigation from being rendered when it has
no height nor width.