feat(layout-modal): add tooltip and enhanced label
Adds a tooltip to the push layout toggle button to clarify that enabling
it will update the layout for all participants based on the presenter's
layout. Additionally, show a different label for moderators that are not
presenter make it clear that the layout to be propagated is the
presenter's layout.