Avoid sending to Hasura subscription that user doesn't have permission.
When receiving an subscription that requires user_logged, the middleware will store the subscription but will not send it.
Once the user log in and the hasura Hole is updated, then the middleware will forward all the subscriptions.
It will avoid the client from receiving a lot of errors when it has a reconnection and automatically re-send all the subscriptions without checking if the client is logged.