fix: replication of layout
Fixes a condition where the layout was only applied when the push layout
of the meeting changes. So changing the layout without changing the push
layout didn't effectively propagate the layout.
Also added a missing flag to the settings yml file:
`public.app.defaultSettings.pushLayout` which is used to initialize the
push layout value in the settings singleton.