Document formatting updates and corrections for 3.0
Improve the formatting of the API parameters table to allow specifying
minimum and maximum values, as well as minimum and maximum lengths for
string values. These get displayed in the "Type" column below the data
type name.

Alter the appearance of required and deprecated parameters to make
required parameters stand out more, and more clearly indicate deprecated
parameters.

I've reformatted some of the existing parameter entries to take
advantage of the new features, and added or corrected descriptions to
several parameters.