feat(core): add new deafened backend state to voice users
LiveKit's audio bridge will not have the traditional "Leave Audio"
approach that we had before, as its connection is supposed to be up and
running at all times. To keep the UI's "Leave Audio" alive, we should
allow users to stop hearing others whenever possible, in the same way
they can mute itself, and handle that state accordingly as a "not in
audio". There are examples of that approach - e.g.: Discord.

Add a new deafened state to the backend so that the implementation above
is possible.