fix(audio): do not log, notify or enforce I/O when joining audio deafened
When connectiong to audio with the deafened flag active, none of the
following procedures need to be triggered as this is not considered a
complete audio join:
  - I/O enforcement
  - UI notification
  - Join log

Return early when audio-manager's onAudioJoin callback is called with
the deafened state set to true. This avoids duplicate and unwarranted
notifications, logs and I/O resets.