fix(livekit): blurry screen sharing due to borked presets
LiveKit's screen sharing presets aren't configurable. The default values
are also not ideal right now, which causes the received stream to be in
a lower quality than it should in specific scenarios due to adaptive
streaming.

Make screen sharing's presets configurable, as well as adjust the
default encoding preset to something that makes for nicer layers
(1080p, 15fps).