fix(screenshare): audio from shared tabs is muted
When using the default bridge (bbb-webrtc-sfu/mediasoup), audio from
a shared tab is muted on the receiver's end. This is caused by a change
in the `view` bridge argument signature that was not applied to the
webrtc-sfu bridge - causing the hasAudio property to be undefined the
media element to be muted.

Adjust bbb-webrtc-sfu's screen sharing bridge so that the options object
and hasAudio property are properly picked up.