fix(livekit): no mic output after joining audio in "listen only" mode
Whenever an user joins as "listen only" (no input device, lock settings
is enforced) and later tries to activate a microphone mid-call, the mic
in question will output no audio. This only affects the LiveKit bridge.
This is due to an uneeded "isListenOnly" flag check in audio-manager's
setSenderTrackEnabled interface.

Remove the "isListenOnly" check in audio-manager's setSenderTrackEnabled
interface as it causes the aforementioned bug and is unnecessary due to:
  - Bridges should handle having no senders on their own
  - The "listen only" flag is transitioning into a virtual, client-only
    state only serves the purpose of tracking whether an input device is
    set. Audio modes aren't locked anymore (transparent listen only,
    LiveKit), so it shouldn't block any internal method other than for
    UI warning purposes.