feat(plugins): add apps gallery extensible area
Adds new extensible area for the apps gallery. Previously, to add an
entry in the apps gallery, a plugin should inject a generic sidekick
content, which means that something would be shown in the sidebar
content. But, as the concept of an app is growing beyond the sidebar
content, this new extensible area allows for injecting an entry in the
apps gallery and can execute any action using the 'onClick' callback.