feat(audio-captions): optionally display it in apps gallery
To address the issue of the action bar becoming overcrowded with
buttons, especially on mobile devices where it can overflow off the
screen, a new 'useAppsGallery' setting has been added to the
settings.yml. When enabled, this setting moves the transcription
controls to a separate panel that can be accessed through the apps
gallery. The default behavior is set to false, preserving the previous
behavior where the transcription settings remain in the action bar.

Additionally it fixes an error where, when the audio captions toggle was
enabled, no language was selected by default. It also restores the
original placement for the CC button to the left of the actions bar.