fix(livekit): paginated cameras sometimes fail to be subscribed
Paginated LK cameras sometimes fail to be subscribed. This is caused by
trailing subscribe/unsubscribe requests happening for a previous page
_after_ a switch to another page was made.

That can happen in a bunch of scenarios, but the most common one is two
separate stream update calls hapenning in a short span that's a) long
enough to circumvent the debounce b) short enough so that the
subscribe/unsubscribe cycle doesn't fully happen. Since subscribed tracks
from LK are not a dependency for stream updates, that desync won't be
corrected.

Add subscribed camera tracks (provided by LK) as dependency to the
updateStreams useEffect. That should correct any state desync regarding
subscribed cameras.