fix(livekit): ICE servers fetch may loop
The ICE servers fetch routine may loop indefintely if the request fails
with an empty dictionary as a result instead of an error. That causes a
client freeze.

Do not reset the ICE servers dictionary if the ICE servers response came
in empty, but did not throw an error. This should prevent the hook from
looping.