chore(livekit): set livekit.audio.unpublishOnMute to true
Unpublish audio tracks on mute by default. This should improve
performance across clients (minimally), livekit-server (minimally)
and egress (significantly).

There might be a drawback with long running sessions or sessions with too
much back-and-forth. Therefore, this may be revised in the future - if
cycling transceivers too much becomes an issue, we'll probably refurbish
this into a "smarter" version.