docs: update the portion about reducing webcam bandwidth
The yq commands were from BigBlueButotn 2.7 (different yq version and syntax)
Also, running them did little to /etc/bigbluebutton/bbb-html.yml unless
the cameraProfiles portion was present and populated.
Therefore I added some optional instructions for how to populate it.
The resulting yml looks fine to me with the exception that
some profiles have a higher bandwidth than desired but are not in the
list of low, medium, high, hd. Seems they are activated for small
sessions only, so likely not a concern bandwidth-wise.