fix(livekit): autoplay modal triggered unnecessarily
Yet another instance of the autoplay modal being triggered
unnecessarily. This time, mostly due to a change to livekit-cli on how
the playback status is defined.

Track the canPlayAudio state from useAudioPlayback and update the local
autoplay state so that the user isn't prompted if the state changes
mid-check. Additionally, add an under-the-hoods play attempt after audio
is connect to force playback state transitioning to success if possible.
If that check fails, the autoplay modal is displayed as before.