Speed up compositing of video in recording processing
This combines a few improvements that I've figured out that are specific
to processing/encoding of video. To summarize:

* Use faster options for intermediate video file encoding
* The overlay filters were a single-thread bottleneck; switch to xstack
* Don't composite videos hidden behind other videos
* Process multiple cuts of video in parallel

I had known about some bottlenecks in the recording processing affecting
the speed when large numbers of individuals were being composited - for
example, a meeting with 20+ webcams. This primarily comes from the use
of the "overlay" filter, which has to be repeated for each video and is
rather inefficient when used in that way.

The use of the overlay filter can be replaced by the xstack filter
(introduced in ffmpeg 4.1), which is much faster at compositing multiple
videos. However, it does not handle overlapping videos well -
overlapping videos can resulting in flickering artifacts.

To solve the overlapping video problem, I added an optimization which
prevents videos that are completely obscured from being composited. For
example, in the current video layout this means that the presentation
area video will be omitted when the screenshare video is covering the
presentation area. This provides an additional speedup from reducing
the number of videos being handled.

Finally, a new feature is added: parallel processing of multiple
recording cuts. Assuming enough system resources are available, doing
the video encode for multiple cuts in parallel can provide a significant
speedup.