Speed up compositing of video in recording processing
This combines a few improvements that I've figured out that are specific
to processing/encoding of video. To summarize:

* Use faster options for intermediate video file encoding
* The overlay filters were a single-thread bottleneck; switch to xstack
* Don't composite videos hidden behind other videos
* Fix a tpad/fps filter bottleneck with long framerate gaps
* Process multiple cuts of video in parallel

The video encoding options were selected based on testing on a variety
of recordings - they provide a reasonable speedup on the h264 encoding
with only minor efficiency loss on the types of video that ffmpeg
handles.

The compositing performed by the "overlay" filter is somewhat
inefficient, especially when large numbers of individual videos are
being composited using a separate overlay filter for each video. The
overlay filter can be replaced by the xstack filter (introduced in
ffmpeg 4.1), which is much faster at compositing multiple videos.
However, it does not handle overlapping videos well - they can result in
flickering artifacts.

To solve the overlapping video problem, I added an optimization which
prevents videos that are completely obscured from being composited. For
example, in the current video layout this means that the presentation
area video will be omitted when the screenshare video is covering the
presentation area. This provides an additional speedup from reducing
the number of videos being handled.

Another bottleneck which can cause some long running (multiple-hour)
recordings to hit extreme slowdowns was related to the order of the tpad
and fps filters. In the case where (partially due to the use of separate
ffmpeg processes for pre-filtering) the input video fps is misdetected
as being extremely high (I saw as high as ~61K fps in one bad file!),
and where the recorded video cuts out or has a very long gap, ffmpeg can
spend several minutes with tpad generating hundreds of thousands of
frames which the fps filter will immediately drop. This is improved by
changing the filter order.

A new feature is added: parallel processing of multiple recording cuts.
Assuming enough system resources are available, doing the video encode
for multiple cuts in parallel can provide a significant speedup. The
parallelism is currently hardcoded as 2 parallel ffmpeg processes;
configuration will be added in a follow-up.