fix(core): increase max authToken size to 50 chars in `user` table
The authToken property in the `user` table has a size limit of 16 chars.
That same property can be as large as the userId (intId) in scenarios
where non-web users (e.g.: dial-in) are involved, since authToken is
spoofed in akka-apps to be the same as that user's intId.
This may cause the insert op for that external user to be failed, which
translates to them being missing in the user list.

Increase the max authToken size to 50 chars in the `user` table, which
matches `userId`.