feat(default-settings): support meeting-scoped settings
This commit allows adding default setting keys to a list of meeting-scoped
settings that won't be persisted between meetings, even when
`userSettingsStorage` is set to 'local'. This is achieved by appending
the meeting ID to the key used for storage.

Also introduces a new settings group for layout-related settings and add
this key to the meeting-scoped list, preventing layout settings - such as
push layout state and selected layout type - from being persisted across
meetings.