Port(html5): Add rollback to video-preview gUM and improve logs for 3.0
PR #21848: fix(webcam): add option to skip initial enumerateDevices
Description:
See: #21746
The client currently does an enumerate -> optional gUM -> enumerate flow
(the last enum only if necessary - i.e.: fingerprinting measures were in
effect in the first enumerate), which is an attempt to save extra
enumerate/gUM calls whenever necessary. The fact that it's breaking gUM
in Firefox seems like a browser bug, but we should work around it.

Add a flag (public.media.skipInitialCamEnumeration) that allows skipping
the initial optimistic enumerateDevices call to try to work around the
aforementioned issue. After this has been tested and validated, we can
port it to 3.0 without a feature flag behind it.

PR #21864: fix(webcam): add rollback to video-preview gUM, improve logs
Description:
Following up on 2b503f4, it'd be good if we had:
  - More info on why getUserMedia fails
  - A rollback procedure in case gUM failures are recoverable.

This commit adds a rollback procedure for video-preview gUM calls whenever
the error type is OverConstrainedError. The retry happens with the
simplest constraint possible.

Additionally:
  - Add more information to video-preview gUM failure logs
  - Add a centralized gUM error log to video-preview's service doGUM
    util
  - Proceed with the gUM routines even if the initial enumeration call
    fails as the user doesn't necessarily need enumeration for cameras
    to be shared