Recording: Remove 'duration' lines from ffmpeg playlist
Although adding these (in theory) improved audio/video sync slightly,
they sometimes cause problems when the video is transcoded to vp9 after
concatenation, if there are multiple short cuts near the start of the
video which cause ffmpeg to identify the framerate of the concatenated
file as a fraction which is close to (but not equal to) the desired
framerate.

Remove these, reverting to the previous behaviour. I have some alternate
ideas for improving audio/video sync that I'm still investigating.