fix(audio): use storage to check if input device was previously set
The check to determine whether a device was previosly set was pulling
directly from AudioManager's inputDeviceID getter. That getter is
updated regardless of whether a device was joined with. This causes an
issue with the skipEchoIfPreviousDevice where the modal might be skipped
even it a device wasn't previously configured for the current
session/domain.

Pull the device ID from the storage instead to guarantee that only those
that were actively used for audio are taken into account. Additionally,
make the check always evaluate to boolean for consistency's sake.