fix(webcam): rejection in getStats call causes whiteboard reload
An unhandled rejection in video-provider's periodic getStats causes
whiteboard to reload in certain scenarios.
The rejection is usually originated when a stale or placeholde ({}) peer
is in the reference dictionary. In those cases, getStats won't be
defined. Since there's no try-catch in the whole getStats block, the
rejection bubbles up to the global rejection catcher and triggers a
whiteboard reload.

Add a try-catch block to video-provider's getStats procedure, as well as
check whether a peer is properly defined before calling getStats.