fix(livekit): outdated participants in selective audio subscription
The selective subscription hooks currently rely on remote participants
only for defining the desired vs current subscription maps.
Those remove participants are filtered to avoid spurious updates, but
the filter is too restrictive and may be impacting our ability to react
to subscription failures, upstream publication state changes et al.

Extend the participant filters with additional events so that the local
refs are more frequently updated. We'll track metrics and log data to see
if this improves the subscription situation.

Additionally: extend the camera bridge track filter as well for the same
reasons above.