fix(audio-captions): empty list of available captions languages
Previously, the list of available caption languages was populated using
the browser's Speech Recognition API. This caused the list to be empty for
users on browsers without this API support, leaving them unable to select
any captions to be displayed.

This change decouples the caption language list from the local Speech
Recognition API. The list is now correctly populated based on the active
transcription languages present in the meeting.

To improve user experience when no one is transcribing, this also adds:
- A tooltip on the captions toggle.
- A message right below the captions toggle when it is disabled, explaining
  that tracks appear as users enable transcription.