fix: reintroduce meta_hack-record-viewer-video
Metadata support was removed from the client and
meta_hack-record-viewer-video was removed alongside it. There are still
some users of that metaparam, so it needs to be reintroduced.

Extend the GetRecordingStatus server-side probe made between webrtc-sfu
and akka-apps so that it responds with media-specific recording consent
states for audio, cameras and screen sharing. Audio and screen sharing
are unimplemented right now (default to true), but cameras work as
follows:
  - Factor meta_hack-record-viewer-video and userdata-bbb_record_video
    into the consent resolution for the requester
  - Respond with the appropriate consent state for cameras via the
    recordUserCameras field.
Actually handling of these consent booleans is done by bbb-webrtc-sfu.