fix(layout): ensure the camera position is not undefined
When changing from one layout to another, there is a merge of the
previous input values with some of new values of the new layout. It
happens that sometimes the previous layout didn't initialized the value
for the camera dock position, leaving it as undefined and causing it to
insert an inconsistent state to the next layout state. So this commit
fixes it by initializing the camera position in all layouts types, to ensure
its value in the layout input is not undefined.