fix(livekit): unable to connect if /stuns maxTtl is undefined
If maxTtl is not defined for the configured ICE servers returned by
the /stuns endpoint, LiveKit won't be able to connect. The hook used to
fetch said ICE servers gets stuck in a loading loop due to the cache
being perpetually invalidated because of the nil TTL.

Assume that if no TTL is set for /stuns, the cache is valid
indefinitely.