fix(chat): ensure message sender is always displayed
Messages sent by plugins may include a `custom` attribute meant to
render a raw message without name or avatar, leaving room for the plugin
to style it afterwards.

However, since plugin messages are triggered by custom window events,
anyone aware of the event format could forge such events and inject chat
messages with arbitrary data. In the UI, these messages would not
display the sender’s name or avatar, making the source unidentifiable to
moderators during a conference.

This commit enforces that messages with the `custom` attribute also
render the sender’s name and avatar. With this change, the purpose of
the `custom` attribute becomes outdated, and we may revisit whether to
maintain it or introduce more subtle ways of identifying the sender,
such as on hover.