feat(layout): initialize context with correct device type
Calls the function to get the device type as soon as the initial state
is declared. There is no need to defer it to the layout initialization,
since it can lead to useless renders, where the layout changes but the
device type is still not set, leading to situations where components
would skip layout updates until the device type is set.