fix(plugins-sdk): compare generic content items by their id
Currently, the diff function is comparing all the object to decide when to add/remove new generic content items,
this has problems when the same generic content (same id) is set by a plugin, causing it to be added and removed wrongly.

This fixes it by comparing the items by their id to decide when to add/remove.