Implement dynamic layout in video recording format
The current video recording format supports multiple layouts depending
on the media shared in the recording, but the layout has to be selected
before processing the recording and a single layout is used for the
entire recording length.

This causes problems with, for example, meetings that use deskshare and
webcams, but do not use presentation. Either the webcams are fullscreen
when deskshare isn't being used but the deskshare covers up the webcams,
or the webcams are beside the deskshare but remain small and at the side
of the recording.

The new system uses the same layouts but instead of selecting the
layout once before generating the video, it selects the layout
separately for each individual cut of the video. This allows, for
example, webcams to be fullscreen when deskshare isn't used, but when
deskshare starts the webcams will become smaller and move to the side.

As part of this work, to allow the presentation video to change size
as the layout changes between cuts, the handling for video sources was
refactored to be generic, allowing a new video source to be made for
presentation video which generates presentation video on demand for each
cut. This also speeds up recording processing and reduces disk space
requirements by removing the separate step to generate the presentation
video before rendering the recording.