Simplify recording state tracking, fix start_time handling
The recording state tracking was unnecessarily complicated in order to
handle the case where a recording start (or stop) was received while the
recording was already started (or stopped). None of the other recording
bothers handling this, and the conditions can be made easier to read by
removing this logic.

To simplify the tracking of last screenshare as content, I'm not reading
it directly from the screenshare_content_events array instead of trying
to keep a separate variable in sync.

Code has been added to handle the special case of the start_time
(recording segment start timestamp) happening after a recording start
event and screenshare as content event have been received. In this case,
it needs to emit an event corresponding to the current screenshare as
content state at the 0 timestamp.